requirements.txt
Flask==2.1.3
Flask_Cors==3.0.10
matplotlib==3.5.2
numpy==1.23.1
Pillow==9.2.0
scikit_image==0.19.3
torch==1.12.0
